practice/base_requirements.txt
PyJWT==2.6.0
itsdangerous==2.1.2
